#174 move type_register.h to dispach
